Day 23: VM works
