+++ 24/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(aoc_2021_24 main.c day24.c)
+add_executable(aoc_2021_24 main.c day24.c virtual_machine.c)
 
 target_link_libraries(aoc_2021_24 file_utils timing pthread)
